Update duchess::Null to return a null pointer
**NOT A WORKING IMPLEMENTATION!!!**

https://github.com/duchess-rs/duchess/issues/161

When running codegen.rs, we prepare our input arguments to pass them to the
Java JNI methods. To do this, we derefence the rust arguments calling `as_jref`
if the argument type is an Object/Reference. The current implmentation for
duchess::Null automatically returns an `Err(NullDeref)` which automatically
fails the call. We replace the the null dereference error with a null pointer
instead and use that as the pointer to the java type. I believe this is the
intended implementation since this is equivalent to Java's null pointer
(0x00000000).

Oddly, even though we record `Ok(null_ref)` from duchess::Null's `as_jref`,
`codegen.rs` prepare_input still fails for an `NullDeref` error right away when
we call `as_jref`.

With this commit, we get the following when we run `
```sh
cd test-crates/duchess-java-tests
cargo build
/Users/nathanng/Developer/duchess/test-crates/target/debug/duchess-java-tests
```

Using `rust-lldb` struggles parsing through the codegenerated files.